+++ config/contacts.php
@@ -1,0 +1,5 @@
+<?php
+
+return [
+    'googleMapsKey' => env('GOOGLE_MAPS_GEOCODING_KEY', '')
+];